backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
motor==3.3.2
beanie==1.23.6

# AI/ML
torch==2.1.0
scikit-learn==1.3.2

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication  ← ADD THESE
PyJWT==2.8.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0

# Utils
python-dotenv==1.0.0
email-validator==2.1.0